Category problem
I am stuck on Categories controller...The function SubmitSelected is not
complete. Please help me to check it out if you guys can find the way.
The thing is I ahve a problem with database. I cannot access into the
user_category...I will try to work on it again tomorrow.